server/requirements.txt
APScheduler==3.10.1
beautifulsoup4==4.12.2
Flask==2.2.2
python-dotenv==1.0.0
requests==2.28.1
selenium==4.9.1
webdriver_manager==3.8.5
gunicorn==20.0.4
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.6.2